Issue #111: Check host OS before opening URL
I've put in a `switch` statement so that the command for opening
the URL is different, depending on which host operating system
we're running on.

The command needs to be different for each OS, and now we'll be
using `open` for Mac OS, `xdg-open` for Linux, and `start` for
Windows.

I've not been able to test this (yet) on Mac OS. I can do that
later on tonight, but it would be helpful if somebody else could
test that, too.

Signed-off-by: Elliot Voris <elliot@voris.me>