Issue #120: Allows the dashboard command to take an optional env flag
If the `--env` flag is specified on the command line, the url will
include that at the end.

If the `--env` flag is not speicified, the command continues to act
as it has. The user will not be prompted to choose an environment,
and the URL will not incoporate any default information.

Signed-off-by: Elliot Voris <elliot@voris.me>