Cleans up access to logger and outputter
Almost all access to logging and output now go through $this->log() or $this->output() inside a command. Failures are triggered by an exception and handled by the runner. Unused logging and output code has been removed. Some references to Terminus::line remain because it was too awkward to remove them.

Addresses #531