Fix connection tests when TERMINUS_HOST is overriden
When there is a local override for TERMINUS_HOST=onebox the environment test fails as it does not account for the change in behavior of the model. This fix is a bit hacky but is hopefully temporary until we rewrite this unit test to isolate it from local configuration etc.