Ask json_decode to produce arrays rather than objects.
See the [manual entry for json_decode](http://php.net/manual/en/function.json-decode.php).

If you do this, then the result you get back will consist of pure associative arrays (and strings). This reduces the need to cast the nested results to (array) everywhere.

I maintained all of the typecasts in the Annotated Commands code, so this is not necessary. You might want to do this to clean up Terminus code, though -- presuming that you are not depending on the results being stdclass objects elsewhere, anyway.