Add connection:get command
`connection:set` throws an exception if the desired `mode` is already set for the environment. This is problematic when using Terminus in a script as the script will fail. The addition of a `connection:get` command will allow the connection mode to be determined so `connection:set` can be conditionally run only if it is needed.