Filter pantheonsite.io domains from interactive choice list.
While the platform supports setting these domains as primary, it seems
highly unlikely that a user would do this intentionally, and more likely
that if it were chosen from a multiple choice list, that it was just
fat-fingered. For better UX, they are therefore excluded.

Users that really mean to do this can do so by typing/passing the full
pantheonsite.io domain non-interactively to the domain:primary:set
command.