[CORE-1576] Commands to set and unset primary domain. (#2011)
* Commands to set and unset primary domain.

WIP: needs tests.

* Tests for setting/resetting primary domain and its commands.

* Rename command to remove primary domain to `domain:primary:reset` for consistency.

* Refactor set_primary_domain workflow maker out of a Command to Models\Environment class.

* CS fix

* Filter pantheonsite.io domains from interactive choice list.

While the platform supports setting these domains as primary, it seems
highly unlikely that a user would do this intentionally, and more likely
that if it were chosen from a multiple choice list, that it was just
fat-fingered. For better UX, they are therefore excluded.

Users that really mean to do this can do so by typing/passing the full
pantheonsite.io domain non-interactively to the domain:primary:set
command.

* Add primary domain to Domain model and domain:list command.

* UX review: Change command to domain:primary:unset, log completion.

* Changes that are clear from @TeslaDethray's requests.

* Rename domain:primary:set to domain:primary:add per consensus of product and technical leadership.

* Best guess at new model refactor request

* Fix failing tests in older minor versions of php 7.0 - 7.3.

It seems substr_compare issues a warning iff the first parameter is
the empty string in these earlier minor releases; manual states it
was fixed in 7.2.18. Since a) it's been fixed and b) empty strings
are not expected real-world input, I've simply removed that input
from the test coverage.

* Behat tests for add and remove primary domain

* Make behat.yml a bit more yaml-compliant (% not allowed as beginning of string)

* Restore Consolidation\\Filter\\Hooks\\FilterHooks to command list

* Changelog for 2.2.0

* Fix stated version in changelog